ChromeDriver.exe process not being released after test pass
Problem - Did not define usage of the @After kekyword which creates a post-process after all tests are ran. Once a new WebDriver was being called, it was never closed.

Even after manually closing the ChromeDriver.exe window in the desktop, an orphaned process was still found on the system.

Fix - Closed WebDriver handle by defining an @After post-process.